refactor: dao를 interface로 추출
