+++ src/main/resources/application.yml
@@ -1,3 +1,13 @@
 handlebars:
   suffix: .html
   enabled: true
+
+spring:
+  datasource:           # datasource 설정
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:mem:testdb
+    username: sa
+    password:
+  h2:
+    console:
+      enabled: true     # h2 console 설정
